moved `generate_notes` to phichain-chart
