+++ phichain-game/src/curve_note_track.rs
@@ -1,8 +1,6 @@
 use crate::GameSet;
 use bevy::prelude::*;
-use num::iter;
-use phichain_chart::beat;
-use phichain_chart::curve_note_track::CurveNoteTrackOptions;
+use phichain_chart::curve_note_track::{generate_notes, CurveNoteTrackOptions};
 use phichain_chart::note::{Note, NoteBundle};
 
 #[derive(Debug, Clone, Component)]
@@ -37,48 +35,6 @@ impl CurveNoteTrack {
             None
         }
     }
-}
-
-/// Generate a note sequence from a note to another note with a [`CurveNoteTrackOptions`] option
-pub fn generate_notes(from: Note, to: Note, options: &CurveNoteTrackOptions) -> Vec<Note> {
-    // make sure from.beat < to.beat
-    let (from, to) = if from.beat < to.beat {
-        (from, to)
-    } else {
-        (to, from)
-    };
-
-    let mirror = from.x > to.x;
-
-    let beats = iter::range_step(
-        from.beat.min(to.beat),
-        from.beat.max(to.beat),
-        beat!(1, options.density),
-    )
-    .collect::<Vec<_>>();
-    let notes = beats
-        .iter()
-        .enumerate()
-        .map(|(i, beat)| {
-            let x = i as f32 / beats.len() as f32;
-            let y = if mirror {
-                1.0 - options.curve.ease(x)
-            } else {
-                options.curve.ease(x)
-            };
-
-            Note::new(
-                options.kind,
-                true,
-                *beat,
-                (from.x - to.x).abs() * y + from.x.min(to.x),
-                1.0,
-            )
-        })
-        .skip(1)
-        .collect::<Vec<_>>();
-
-    notes
 }
 
 pub struct CurveNoteTrackPlugin;
